Add all payment method types
