+++ client/extensions/woocommerce/app/settings/payments/payments-off-site.js
@@ -1,105 +1,22 @@
 /**
  * External dependencies
  */
-import React, { Component, PropTypes } from 'react';
-import { bindActionCreators } from 'redux';
-import { connect } from 'react-redux';
+import React, { Component } from 'react';
 import { localize } from 'i18n-calypso';
 
 /**
  * Internal dependencies
  */
-import {
-	changePaymentMethodField,
-	closeEditingPaymentMethod,
-	openPaymentMethodForEdit,
-} from 'woocommerce/state/ui/payments/methods/actions';
-import { errorNotice, successNotice } from 'state/notices/actions';
 import ExtendedHeader from 'woocommerce/components/extended-header';
-import { fetchPaymentMethods, savePaymentMethod } from 'woocommerce/state/sites/payment-methods/actions';
-import { getCurrentlyEditingPaymentMethod, getPaymentMethodsGroup } from 'woocommerce/state/ui/payments/methods/selectors';
-import { getSelectedSiteId } from 'state/ui/selectors';
-import List from 'woocommerce/components/list/list';
-import ListHeader from 'woocommerce/components/list/list-header';
-import ListItemField from 'woocommerce/components/list/list-item-field';
-import PaymentMethodEdit from './payment-method-edit';
-import PaymentMethodItem from './payment-method-item';
+import PaymentMethodList from './payment-method-list';
 
-class SettingsPaymentsOffSite extends Component {
-	static propTypes = {
-		closeEditingPaymentMethod: PropTypes.func.isRequired,
-		currentlyEditingMethod: PropTypes.shape( {
-			id: PropTypes.string,
-		} ),
-		fetchPaymentMethods: PropTypes.func.isRequired,
-		openPaymentMethodForEdit: PropTypes.func.isRequired,
-		paymentMethods: PropTypes.array.isRequired,
-		savePaymentMethod: PropTypes.func.isRequired,
-		siteId: PropTypes.number.isRequired,
-	};
-
-	componentDidMount() {
-		this.props.fetchPaymentMethods();
-	}
-
-	onCancel = ( method ) => {
-		this.props.closeEditingPaymentMethod( this.props.siteId, method.id );
-	}
-
-	onEdit = ( method ) => {
-		const { siteId } = this.props;
-		this.props.openPaymentMethodForEdit( siteId, method.id );
-	}
-
-	onEditField = ( field, value ) => {
-		this.props.changePaymentMethodField( this.props.siteId, field, value );
-	}
-
-	onSave = ( method ) => {
-		const { siteId, translate } = this.props;
-
-		const successAction = () => {
-			this.props.closeEditingPaymentMethod( siteId, method.id );
-			return successNotice(
-				translate( 'Payment method successfully saved.' ),
-				{ duration: 4000 }
-			);
-		};
-
-		const errorAction = () => {
-			return errorNotice(
-				translate( 'There was a problem saving the payment method. Please try again.' )
-			);
-		};
-
-		this.props.savePaymentMethod( siteId, method, successAction, errorAction );
-	}
-
-	renderMethodItem = ( method ) => {
-		const currentlyEditingId = this.props.currentlyEditingMethod &&
-			this.props.currentlyEditingMethod.id;
-		return (
-			<div key={ method.title }>
-				<PaymentMethodItem
-					currentlyEditingId={ currentlyEditingId }
-					method={ method }
-					onCancel={ this.onCancel }
-					onEdit={ this.onEdit } />
-				{ currentlyEditingId === method.id && (
-					<PaymentMethodEdit
-						method={ this.props.currentlyEditingMethod }
-						onEditField={ this.onEditField }
-						onSave={ this.onSave } />
-				) }
-			</div>
-		);
-	}
+class SettingsPaymentsOnSite extends Component {
 
 	render() {
-		const { translate, paymentMethods } = this.props;
+		const { translate } = this.props;
 
 		return (
-			<div className="payments__off-site-container">
+			<div className="payments__type-container">
 				<ExtendedHeader
 					label={ translate( 'Off-site credit card payment methods' ) }
 					description={
@@ -109,48 +26,10 @@ class SettingsPaymentsOffSite extends Component {
 							'information'
 						)
 					} />
-					<List>
-						<ListHeader>
-							<ListItemField className="payments__methods-column-method">
-								{ translate( 'Method' ) }
-							</ListItemField>
-							<ListItemField className="payments__methods-column-fees">
-								{ translate( 'Fees' ) }
-							</ListItemField>
-							<ListItemField className="payments__methods-column-settings">
-							</ListItemField>
-						</ListHeader>
-						{ paymentMethods && paymentMethods.map( this.renderMethodItem ) }
-					</List>
+					<PaymentMethodList methodType="off-site" />
 			</div>
 		);
 	}
 }
 
-function mapStateToProps( state ) {
-	const currentlyEditingMethod = getCurrentlyEditingPaymentMethod( state );
-	const paymentMethods = getPaymentMethodsGroup( state, 'off-site' );
-	const siteId = getSelectedSiteId( state );
-	return {
-		currentlyEditingMethod,
-		paymentMethods,
-		siteId,
-	};
-}
-
-function mapDispatchToProps( dispatch ) {
-	return bindActionCreators(
-		{
-			changePaymentMethodField,
-			closeEditingPaymentMethod,
-			fetchPaymentMethods,
-			openPaymentMethodForEdit,
-			savePaymentMethod,
-		},
-		dispatch
-	);
-}
-
-export default localize(
-	connect( mapStateToProps, mapDispatchToProps )( SettingsPaymentsOffSite )
-);
+export default localize( SettingsPaymentsOnSite );
