+++ client/extensions/woocommerce/app/settings/payments/payments-offline.js
@@ -7,15 +7,15 @@ import { localize } from 'i18n-calypso';
 /**
  * Internal dependencies
  */
-import Card from 'components/card';
 import ExtendedHeader from 'woocommerce/components/extended-header';
+import PaymentMethodList from './payment-method-list';
 
 class SettingsPaymentsOffline extends Component {
 
 	render() {
 		const { translate } = this.props;
 		return (
-			<div>
+			<div className="payments__type-container">
 				<ExtendedHeader
 					label={ translate( 'Offline payment methods' ) }
 					description={
@@ -24,7 +24,7 @@ class SettingsPaymentsOffline extends Component {
 							'transfer, check or cash on delivery.'
 						)
 					} />
-				<Card></Card>
+				<PaymentMethodList methodType="offline" />
 			</div>
 		);
 	}
