+++ client/extensions/woocommerce/app/settings/payments/payments-on-site.js
@@ -7,29 +7,29 @@ import { localize } from 'i18n-calypso';
 /**
  * Internal dependencies
  */
-import Card from 'components/card';
 import ExtendedHeader from 'woocommerce/components/extended-header';
+import PaymentMethodList from './payment-method-list';
 
 class SettingsPaymentsOnSite extends Component {
 
 	render() {
 		const { translate } = this.props;
+
 		return (
-			<div>
+			<div className="payments__type-container">
 				<ExtendedHeader
 					label={ translate( 'On-site credit card payment methods' ) }
 					description={
 						translate(
-							'On-site methods provide a seamless experience by keeping the ' +
-							'customer on your site to enter their credit card details and ' +
-							'complete checkout. More information'
+							'On-site payment methods involve sending the customer to a ' +
+							'third party web site to complete payment, like PayPal. More ' +
+							'information'
 						)
 					} />
-				<Card></Card>
+					<PaymentMethodList methodType="on-site" />
 			</div>
 		);
 	}
-
 }
 
 export default localize( SettingsPaymentsOnSite );
